+++ utils/count_hex_digit.c
@@ -6,11 +6,11 @@
 /*   By: abe21453@cs.saisoncard.co.jp <abe21453@    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/06/06 00:46:51 by abeyuuta          #+#    #+#             */
-/*   Updated: 2023/06/07 16:19:18 by abe21453@cs      ###   ########.fr       */
+/*   Updated: 2023/06/09 01:32:12 by abe21453@cs      ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int	count_hex_digit(unsigned int num)
+int	count_hex_digit(unsigned long long num)
 {
 	int	count;
